+++ nominal/core/connection.py
@@ -2,9 +2,10 @@ from __future__ import annotations
 
 import itertools
 import logging
+import warnings
 from dataclasses import dataclass, field
 from datetime import timedelta
-from typing import Iterable, Mapping, Protocol, Sequence, cast
+from typing import Iterable, Mapping, Protocol, Sequence
 
 from nominal_api import (
     datasource_api,
@@ -17,10 +18,10 @@ from nominal_api import (
 )
 
 from nominal.core._clientsbunch import HasAuthHeader, ProtoWriteService
-from nominal.core._utils import HasRid, _to_api_batch_key
+from nominal.core._utils import HasRid
+from nominal.core.batch_processor import process_batch_legacy
 from nominal.core.channel import Channel
-from nominal.core.stream import BatchItem, WriteStream
-from nominal.ts import _SecondsNanos
+from nominal.core.stream import WriteStream
 
 
 @dataclass(frozen=True)
@@ -144,8 +145,6 @@ class Connection(HasRid):
         """get_nominal_write_stream is deprecated and will be removed in a future version,
         use get_write_stream instead.
         """
-        import warnings
-
         warnings.warn(
             "get_nominal_write_stream is deprecated and will be removed in a future version,"
             "use get_write_stream instead.",
@@ -185,18 +184,23 @@ class Connection(HasRid):
             ```
 
         """
-        if self._nominal_data_source_rid is None:
-            raise ValueError("Writing not implemented for this connection type")
-
         if not use_protos:
+            warnings.warn(
+                "Using the batch processor without protos is deprecated and will be removed in a future version, "
+                "use get_write_stream(use_protos=True) instead.",
+                UserWarning,
+                stacklevel=2,
+            )
             return WriteStream.create(
                 batch_size,
                 max_wait,
-                self._process_batch,
+                lambda batch: process_batch_legacy(
+                    batch, self._nominal_data_source_rid, self._clients.auth_header, self._clients.storage_writer
+                ),
             )
 
         try:
-            from nominal.core.batch_processor import process_batch
+            from nominal.core.batch_processor_proto import process_batch
         except ImportError:
             raise ImportError("nominal-api-protos is required to use get_write_stream")
 
@@ -209,53 +213,6 @@ class Connection(HasRid):
                 auth_header=self._clients.auth_header,
                 proto_write=self._clients.proto_write,
             ),
-        )
-
-    def _process_batch(self, batch: Sequence[BatchItem]) -> None:
-        api_batched = itertools.groupby(sorted(batch, key=_to_api_batch_key), key=_to_api_batch_key)
-
-        if self._nominal_data_source_rid is None:
-            raise ValueError("Writing not implemented for this connection type")
-
-        api_batches = [list(api_batch) for _, api_batch in api_batched]
-
-        def make_points(api_batch: Sequence[BatchItem]) -> storage_writer_api.Points:
-            if isinstance(api_batch[0].value, str):
-                return storage_writer_api.Points(
-                    string=[
-                        storage_writer_api.StringPoint(
-                            timestamp=_SecondsNanos.from_flexible(item.timestamp).to_api(),
-                            value=cast(str, item.value),
-                        )
-                        for item in api_batch
-                    ]
-                )
-            if isinstance(api_batch[0].value, float):
-                return storage_writer_api.Points(
-                    double=[
-                        storage_writer_api.DoublePoint(
-                            timestamp=_SecondsNanos.from_flexible(item.timestamp).to_api(),
-                            value=cast(float, item.value),
-                        )
-                        for item in api_batch
-                    ]
-                )
-            raise ValueError("only float and string are supported types for value")
-
-        request = storage_writer_api.WriteBatchesRequest(
-            data_source_rid=self._nominal_data_source_rid,
-            batches=[
-                storage_writer_api.RecordsBatch(
-                    channel=api_batch[0].channel_name,
-                    points=make_points(api_batch),
-                    tags=api_batch[0].tags or {},
-                )
-                for api_batch in api_batches
-            ],
-        )
-        self._clients.storage_writer.write_batches(
-            self._clients.auth_header,
-            request,
         )
 
     def archive(self) -> None:
